+++ src/components/DeviceDetails/index.test.tsx
@@ -15,15 +15,17 @@ jest.mock('react-router-dom', () => ({
 }));
 
 describe('DeviceDetails', () => {
-  it('should render device', () => {
+  it.only('should render device', () => {
     const props:Props = {
       device: {
         id: 42,
         soort: 'Luchtkwaliteit',
         category: 'Sensor',
-        privacy: 'privacy',
         contact: 'iothings',
         organisation: 'GGD Amsterdam',
+        properties: {
+          Privacyverklaring: 'https://www.amsterdam.nl/privacy/specifieke/privacyverklaringen-b/vaarwegbeheer/',
+        },
       },
       selectedLayer: 'devices',
       onDeviceDetailsClose: jest.fn(),
@@ -34,9 +36,9 @@ describe('DeviceDetails', () => {
     );
 
     expect(queryByText('Apparaat')).toBeInTheDocument();
-    expect(queryByText(props.device?.soort)).toBeInTheDocument();
+    expect(queryByText(props.device.soort)).toBeInTheDocument();
     expect(queryByText(props.device.category)).toBeInTheDocument();
-    expect(queryByText(props.device.privacy)).toBeInTheDocument();
+    expect(queryByText(props.device.properties.Privacyverklaring)).toBeInTheDocument();
     expect(queryByText(props.device.organisation)).toBeInTheDocument();
 
     expect(queryByTestId('closeButton')).toBeInTheDocument();
